+++ service/controller/resource/cleanuptcnpiamroles/create.go
@@ -1,4 +1,4 @@
-package cleanupiamroles
+package cleanuptcnpiamroles
 
 import (
 	"context"
